+++ src/django_pg_migration_tools/operations.py
@@ -6,6 +6,7 @@ from django.contrib.postgres import operations as psql_operations
 from django.db import migrations, models
 from django.db.backends.base import schema as base_schema
 from django.db.migrations.operations import base as base_operations
+from django.db.migrations.operations import models as operation_models
 
 
 class BaseIndexOperation(
@@ -35,6 +36,7 @@ class BaseIndexOperation(
         from_state: migrations.state.ProjectState,
         to_state: migrations.state.ProjectState,
         index: models.Index,
+        unique: bool,
         model: type[models.Model],
     ) -> None:
         self._ensure_not_in_transaction(schema_editor)
@@ -50,6 +52,9 @@ class BaseIndexOperation(
         index_sql = index_sql.replace(
             "CREATE INDEX CONCURRENTLY", "CREATE INDEX CONCURRENTLY IF NOT EXISTS"
         )
+        if unique:
+            index_sql = index_sql.replace("CREATE INDEX", "CREATE UNIQUE INDEX")
+
         schema_editor.execute(index_sql)
         self._ensure_original_lock_timeout_is_reset(schema_editor)
 
@@ -157,6 +162,7 @@ class SaferAddIndexConcurrently(
             to_state=to_state,
             index=self.index,
             model=model,
+            unique=False,
         )
 
     def database_backwards(
@@ -226,5 +232,171 @@ class SaferRemoveIndexConcurrently(
             to_state=to_state,
             index=index,
             model=model,
-            operation=self,
+            unique=False,
+        )
+
+
+class ConstraintOperationError(Exception):
+    pass
+
+
+class ConstraintAlreadyExists(ConstraintOperationError):
+    pass
+
+
+class SaferAddUniqueConstraint(BaseIndexOperation, operation_models.AddConstraint):
+    model_name: str
+    constraint: models.UniqueConstraint
+    raise_if_exists: bool
+
+    _CHECK_EXISTING_CONSTRAINT_QUERY = dedent("""
+        SELECT conname
+        FROM pg_catalog.pg_constraint
+        WHERE conname = %(constraint_name)s;
+    """)
+
+    def __init__(
+        self,
+        model_name: str,
+        constraint: models.UniqueConstraint,
+        raise_if_exists: bool = True,
+    ) -> None:
+        self.raise_if_exists = raise_if_exists
+        self.constraint = constraint
+
+        # Perform a basic input-validation at initialisation time rather than
+        # at migration time ("database_forwards"/"database_backwards") so that
+        # all the errors we __can__ raise at initialisation time are raised
+        # when the programmer is still coding rather than when the migration
+        # effectively runs. This is the "Django way" of doing it, as other
+        # operation classes take this same approach. See "AddIndex" operation.
+        self._validate()
+
+        super().__init__(model_name, constraint)
+
+    def database_forwards(
+        self,
+        app_label: str,
+        schema_editor: base_schema.BaseDatabaseSchemaEditor,
+        from_state: migrations.state.ProjectState,
+        to_state: migrations.state.ProjectState,
+    ) -> None:
+        if not self._can_migrate(app_label, schema_editor, from_state):
+            return
+
+        if not self._can_create_constraint(schema_editor):
+            return
+
+        index = self._get_index_for_constraint()
+        model = to_state.apps.get_model(app_label, self.model_name)
+        self.safer_create_index(
+            app_label=app_label,
+            schema_editor=schema_editor,
+            from_state=from_state,
+            to_state=to_state,
+            index=index,
+            model=model,
+            unique=True,
+        )
+
+        # Django doesn't have a handy flag "using=..." so we need to alter the
+        # SQL statement manually. We go from a SQL that looks like this:
+        #
+        # - ALTER TABLE "table" ADD CONSTRAINT "constraint" UNIQUE ("field")
+        #
+        # Into a SQL that looks like:
+        #
+        # - ALTER TABLE "table" ADD CONSTRAINT "constraint" UNIQUE USING INDEX "idx"
+        base_sql = str(self.constraint.create_sql(model, schema_editor))
+        alter_table_sql = base_sql.split(" UNIQUE")[0]
+        sql = f'{alter_table_sql} UNIQUE USING INDEX "{index.name}"'
+
+        # Now we can execute the schema change. We have lock timeouts back in
+        # place after creating the index that would prevent this operation from
+        # running for too long if it's blocked by another query. Otherwise,
+        # this operation should actually be quite fast - if it's not blocked -
+        # since we have created the unique index in the previous step.
+        return schema_editor.execute(sql)
+
+    def database_backwards(
+        self,
+        app_label: str,
+        schema_editor: base_schema.BaseDatabaseSchemaEditor,
+        from_state: migrations.state.ProjectState,
+        to_state: migrations.state.ProjectState,
+    ) -> None:
+        if not self._can_migrate(app_label, schema_editor, from_state):
+            return
+
+        if not self._constraint_exists(schema_editor):
+            # Nothing to delete.
+            return
+
+        # Execute the DDL that will drop the constraint.
+        super().database_backwards(
+            app_label=app_label,
+            schema_editor=schema_editor,
+            from_state=from_state,
+            to_state=to_state,
+        )
+
+    def _validate(self) -> None:
+        if not isinstance(self.constraint, models.UniqueConstraint):
+            raise ValueError(
+                "SaferAddUniqueConstraint only supports the UniqueConstraint class"
+            )
+
+    def _constraint_exists(
+        self,
+        schema_editor: base_schema.BaseDatabaseSchemaEditor,
+    ) -> bool:
+        cursor = schema_editor.connection.cursor()
+        cursor.execute(
+            self._CHECK_EXISTING_CONSTRAINT_QUERY,
+            {"constraint_name": self.constraint.name},
+        )
+        return bool(cursor.fetchone())
+
+    def _can_create_constraint(
+        self,
+        schema_editor: base_schema.BaseDatabaseSchemaEditor,
+    ) -> bool:
+        constraint_exists = self._constraint_exists(schema_editor)
+        if self.raise_if_exists and constraint_exists:
+            raise ConstraintAlreadyExists(
+                f"Cannot create a constraint with the name "
+                f"{self.constraint.name} because a constraint of the same "
+                f"name already exists. If you want to skip this operation "
+                f"when the constraint already exists, run the operation "
+                f"with the flag `skip_if_exists=True`."
+            )
+        # We can't re-create a constraint that already exists because the
+        # ALTER TABLE ... ADD CONSTRAINT is not idempotent.
+        return not constraint_exists
+
+    def _get_index_for_constraint(self) -> models.Index:
+        return models.Index(
+            *self.constraint.expressions,
+            fields=self.constraint.fields,
+            name=self.constraint.name,
+            condition=self.constraint.condition,
+            opclasses=self.constraint.opclasses,  # type: ignore[attr-defined]
+        )
+
+    def _can_migrate(
+        self,
+        app_label: str,
+        schema_editor: base_schema.BaseDatabaseSchemaEditor,
+        from_state: migrations.state.ProjectState,
+    ) -> bool:
+        model = from_state.apps.get_model(app_label, self.model_name)
+        return bool(self.allow_migrate_model(schema_editor.connection.alias, model))
+
+    def describe(self) -> str:
+        return (
+            f"Concurrently adds a UNIQUE index {self.constraint.name} on model "
+            f"{self.model_name} on field(s) {self.constraint.fields} if the "
+            f"index does not exist. Then, adds the constraint using the just-created "
+            f"index. NOTE: Using django_pg_migration_tools SaferAddUniqueConstraint "
+            f"operation."
         )